docs: document public items in iroh-gossip
also renames mod.rs files to modname.rs
