+++ iroh-gossip/src/lib.rs
@@ -1,4 +1,7 @@
-pub mod proto;
+//! Broadcast messages to peers subscribed to a topic
+
+#![deny(missing_docs, rustdoc::broken_intra_doc_links)]
 
 #[cfg(feature = "net")]
 pub mod net;
+pub mod proto;
